Apply discussion comments
