+++ lib/LanguageServerRename/Model/Exception/EmptyNewName.php
@@ -4,8 +4,8 @@ namespace Phpactor\Extension\LanguageServerRename\Model\Exception;
 
 class EmptyNewName extends \Exception
 {
-    public function __construct(string $message = "", int $code = 0, \Throwable $previous = null)
+    public function __construct(string $message = "")
     {
-        parent::__construct($message, $code, $previous);
+        parent::__construct($message);
     }
 }
